Calculo de Pobreza e Indigencia

Función para calcular la pobreza e indigencia siguiendo la metodología de línea.

calculate_poverty(base, basket, print_summary = TRUE)

Arguments

base
Base individual de uno o más períodos

basket
basket basica alimentaria y total, con la siguiente estructura: # A tibble: N x 5 region periodo CBA CBT codigo <chr> <chr> <dbl> <dbl> <dbl> 1 Cuyo 2016.3 1509. 3872. 42 2 Cuyo 2016.4 1570. 4030. 42 3 GBA 2016.3 1684. 4053. 1 . ... ..... ..... .... ...

print_summary
TRUE/FALSE, opcion para imprimir las tasas de pobreza e indigencia

Details

#'disclaimer: El script no es un producto oficial de INDEC.

Examples

#'
bases <- dplyr::bind_rows(toybase_individual_2016_03,toybase_individual_2016_04)

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

base_pobreza <- calculate_poverty(base = bases, basket = canastas_reg_example,print_summary=TRUE)

#> # A tibble: 2 x 4
#> # Groups:   ANO4 [1]
#>    ANO4 TRIMESTRE Tasa_pobreza Tasa_indigencia
#>   <int>     <int>        <dbl>           <dbl>
#> 1  2016         3       0.0296          0.0136
#> 2  2016         4       0.0324          0.0122