Seleccion aleatoria de casos de la base 2016 trimestre 3 para la base individual

Seleccion aleatoria de casos de la base 2016 trimestre 3 para la base individual

toybase_individual_2016_03

Formato

Un data frame con 2000 filas y 177 variables

CODUSU

'Codigo para distinguir VIVIENDAS, permite aparearlas con Hogares y Personas. Ademas permite hacer el seguimiento a traves de los trimestres'

NRO_HOGAR

' Codigo para distinguir HOGARES, permite aparearlos con Personas'

COMPONENTE

'Numero de componente: Nro de orden que se asigna a las personas que conforman cada hogar de la vivienda'

H15

'Entrevista individual realizada'

ANO4

'Anio de relevamiento'

TRIMESTRE

'Ventana de Observacion'

REGION

'Codigo de Region'

MAS_500

'Aglomerados segun tamanio'

AGLOMERADO

'Codigo de Aglomerado'

PONDERA

'Ponderacion'

CH03

'Relacion de parentesco'

CH04

'Sexo'

CH05

'Fecha de nacimiento (dia, mes y anio)'

CH06

'Edad en anios cumplidos'

CH07

'Actuamente esta...'

CH08

'Tipo de cobertura medica'

CH09

'Sabe leer y escribir?'

CH10

'Asiste o asistio a algun establecimiento educativo? (colegio, escuela, universidad)'

CH11

'Ese establecimiento es...'

CH12

'Cual es el nivel mas alto que cursa o curso?'

CH13

'Finalizo ese nivel?'

CH14

'Cual fue el ultimo anio que aprobo?'

CH15

'Donde nacio?'

CH15_COD

'Codigo de provincia / pais / territorio'

CH16

'Donde vivia hace 5 anios?'

CH16_COD

'Especificar'

NIVEL_ED

'Nivel Educativo'

ESTADO

'Condicion de actividad'

CAT_OCUP

'Categoria ocupacional'

CAT_INAC

'Categoria de inactividad'

IMPUTA

'Indica los casos que han sido imputados'

PP02C1

'De Que manera estuvo buscando trabajo? - Hizo contactos, entrevistas'

PP02C2

'De Que manera estuvo buscando trabajo? - Mando curriculum, puso o contesto avisos'

PP02C3

'De Que manera estuvo buscando trabajo? - Se presento en establecimientos'

PP02C4

'De Que manera estuvo buscando trabajo? - Hizo algo para ponerse por su cuenta'

PP02C5

'De Que manera estuvo buscando trabajo? - Puso carteles en negocios, pregunto en el barrio'

PP02C6

'De Que manera estuvo buscando trabajo? - Consulto a parientes, amigos'

PP02C7

'De Que manera estuvo buscando trabajo? - Se anioto en bolsas, listas, planes de empleo, agencias, contratistas o alguien le estpa buscando trabajo'

PP02C8

'De Que manera estuvo buscando trabajo? - Busco trabajo de otra forma activa'

PP02E

'Durante esos 30 dias no busCo trabajo, porque...'

PP02H

'En los ultimos 12 meses, busco trabajo en algun momento?'

PP02I

'En los ultimos 12 meses, trabajo en algun momento?'

PP03C

'La semana pasada, tenia...'

PP03D

'Cantidad de ocupaciones.'

PP3E_TOT

'Total de horas que trabajo en la semana en la ocupacion principal'

PP3F_TOT

'Total de horas que trabajo en la semana en otras ocupaciones'

PP03G

'La semana pasada, queria trabajar mas horas?'

PP03H

'Si hubiera conseguido mas horas...'

PP03I

'En los ultimos 30 dias, busco trabajar mas horas?'

PP03J

'Aparte de este/os trabajo/s, estuvo buscando algun empleo / ocupacion / actividad?'

INTENSI

'INTENSI'

PP04A

'El negocio / empresa / institucion / actividad en la que trabaja es (se refiere al que trabaja mas horas semanales)...'

PP04B_COD

'A que se dedica o produce el negocio / empresa / institucion?'

PP04B1

'Si presta servicio domestico en hogares particulares'

PP04B2

'En cuantas casas trabaja?'

PP04B3_MES

'Cuanto tiempo hace que trabaja alli? - MES'

PP04B3_ANO

'Cuanto tiempo hace que trabaja alli? - ANIO'

PP04B3_DIA

'Cuanto tiempo hace que trabaja alli? - DiAS'

PP04C

'Cuantas personas, incluido... trabajan alli en total?'

PP04C99

'Tamanio del establecimiento'

PP04D_COD

'Codigo de ocupacion'

PP04G

'Donde realiza principalmente sus tareas?'

PP05B2_MES

'Cuanto tiempo hace que trabaja en ese emPPleo en forma continua? - MES'

PP05B2_ANO

'Cuanto tiempo hace que trabaja en ese emPPleo en forma continua? - ANIO'

PP05B2_DIA

'Cuanto tiempo hace que trabaja en ese emPPleo en forma continua? - DIA'

PP05C_1

'En ese negocio / empresa / actividad, tiene... maquinarias / equipos?'

PP05C_2

'En ese negocio / empresa / actividad, tiene... local (incluye kiosko, puesto fijo)?'

PP05C_3

'En ese negocio / empresa / actividad, tiene... vehiculo?'

PP05E

'Para la actividad del negocio, en los ultimos 3 meses, tuvo que gastar en la compra de materias primas, productos, pagar servicios u otros gastos?'

PP05F

'Ese negocio / empresa / actividad, trabaja habitualmente para...'

PP05H

'Durante cuanto tiempo ha estado trabajando en ese empleo de forma continua? (con interrupciones laborales no mayores a 15 dias)'

PP06A

'En ese negocio / empresa / actividad tiene socios o familiares asociados?'

PP06C

'Monto de patrones y cuenta propia sin socios'

PP06D

'Monto de patrones y cuenta propia con socios'

PP06E

'Ese negocio / empresa / actividad...'

PP06H

'Es una actividad familiar?'

PP07A

'Cuanto tiempo hace que esta trabajando en ese empleo en forma continua? (sin interrupccion de la relacion laboral en la misma empresa / negocio / institucion)'

PP07C

'Ese empleo tiene tiempo de finalizacion?'

PP07D

'Por cuanto tiempo es ese trabajo?'

PP07E

'Ese trabajo es...?'

PP07F1

'En ese trabajo le dan... de comer gratis en el lugar de trabajo'

PP07F2

'En ese trabajo le dan... vivienda?'

PP07F3

'En ese trabajo le dan... algun producto o mercaderia?'

PP07F4

'En ese trabajo le dan... aglun otro beneficio? (automovil, telefono celular, pasajes, etc.)'

PP07F5

'En ese trabajo le dan... no recibe ninguno?'

PP07G1

'En ese trabajo tiene... vacaciones pagas?'

PP07G2

'En ese trabajo tiene... aguinaldo?'

PP07G3

'En ese trabajo tiene... dias pagos por enfermedad?'

PP07G4

'En ese trabajo tiene... obra social?'

PP07G_59

'En ese trabajo tiene... no tiene ninguno?'

PP07H

'Por ese trabajo tiene descuento jubilatorio?'

PP07I

'Aporta por si mismo a algun sistema jubilatorio?'

PP07J

'El turno habitual de trabajo es...?'

PP07K

'Cuando cobra...'

PP08D1

'Cuanto cobro por ese mes por esos conceptos? Monto total de sueldos/jornales, salario familiar, horas extras, otras bonificaciones habituales y tickets, vales o similares'

PP08D4

'Por el mes de...(mes)....cobro... Monto percibido en tickets.'

PP08F1

'Cuanto cobro por ese mes de ..(mes)... Monto en pesos cobrado por Comision por venta/produccion.'

PP08F2

'Cuanto cobro por ese mes de ..(mes)... Monto en pesos cobrado por Propinas.'

PP08J1

'Cuanto cobro por ese mes de ..(mes)... Monto aguinaldo.'

PP08J2

'Cuanto cobro por ese mes de ..(mes)... Monto otras bonificaciones no habituales.'

PP08J3

'Cuanto cobro por ese mes de ..(mes)... Monto retroactivos.'

PP09A

'SOLO OCUPADOS DE: CIUDAD DE BUENOS AIRES. Partidos del GBA En su ocupacion (o en la de mas horas)trabaja....en ....'

PP09A_ESP

'Especificar: contiene la descripcion de otro lugar.'

PP09B

'En su ocupacion (o en la de mas horas), trabaja... en esta ciudad?'

PP09C

'Donde trabaja...?'

PP09C_ESP

'Descripcion de otro lugar (segun pregunta PP09C).'

PP10A

'Cuanto hace que esta buscando trabajo..?'

PP10C

'Durante ese tiempo hizo algun trabajo / changa'

PP10D

'Has trabajdo alguna vez?'

PP10E

'Cuanto tiempo hace que termino su ultimo trabajo / changa?'

PP11A

'El negocio / empresa / institucion / actividad en la que trabaja era...?'

PP11B_COD

'A que se dedicaba o que producia ese negocio/empresa/institucion?'

PP11B1

'Si prestaba servicios domesticos en hogares particulares,'

PP11B2_DIA

'Cuanto tiempo trabajo alli? - DIAS'

PP11B2_MES

'Cuanto tiempo trabajo alli? - MESES'

PP11B2_ANO

'Cuanto tiempo trabajo alli? - ANIOS'

PP11C

'Cuantas personas, inluido ... trabajan alli en total?'

PP11C99

'Ns./Nr. cuantas personas trabajan alli en total'

PP11D_COD

'Como se llamaba la ocupacion que tenia?'

PP11G_ANO

'Cuanto tiempo seguido estuvo trabajando en ese lugar? - ANIO'

PP11G_MES

'Cuanto tiempo seguido estuvo trabajando en ese lugar? - MES'

PP11G_DIA

'Cuanto tiempo seguido estuvo trabajando en ese lugar? - DIA'

PP11L

'Cual fue la razon principal por la que dejo esa actividad?'

PP11L1

'Ese trabajo era...'

PP11M

'Ese trabajo era...'

PP11N

'En ese trabajo lehacian descuento jubilatorio?'

PP11O

'Cual fue la razon principal por la que dejo ese trabajo?'

PP11P

'Cerro la empresa?'

PP11Q

'Fue la unica persona que quedo sin trabajo?'

PP11R

'Le enviaron telegrama?'

PP11S

'Le pagaron indemnizacion?'

PP11T

'Esta cobrando seguro de desempleo?'

P21

'MONTO DE INGRESO DE LA OCUPACIoN PRINCIPAL.'

DECOCUR

'Nro de decil de ingreso de la ocupacion principal del TOTAL EPH.'

IDECOCUR

'Nro de decil de ingreso de la ocupacion principal del INTERIOR EPH'

RDECOCUR

'Nro de decil de ingreso de la ocupacion principal de la REGION'

GDECOCUR

'Nro de decil de ingreso de la ocupacion principal del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES'

PDECOCUR

'Nro de decil de ingreso de la ocupacion principal del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES'

ADECOCUR

'Nro de decil de ingreso de la ocupacion principal del AGLOMERADO.'

PONDIIO

'Ponderador del ingreso de la ocupacion principal.'

TOT_P12

'MONTO DE INGRESO DE OTRAS OCUPACIONES. (Incluye: ocupacion secundaria, ocupacion previa a la semana de referencia, deudas/retroactivos por ocupaciones anteriores al mes de referencia,

etc).',
P47T

'MONTO DE INGRESO TOTAL INDIVIDUAL(sumatoria ingresos laborales y no laborales).'

DECINDR

'Nro de decil de ingreso total individual del TOTAL EPH.'

IDECINDR

'Nro de decil de ingreso total individual del INTERIOR EPH.'

RDECINDR

'Nro de decil de ingreso total individual de la REGION.'

GDECINDR

'Nro de decil de ingreso total individual del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES.'

PDECINDR

'Nro de decil de ingreso total individual del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES.'

ADECINDR

'Nro de decil de ingreso total individual del AGLOMERADO.'

PONDII

'Ponderador para ingreso total individual.'

V2_M

'Monto del ingreso por JUBILACION O PENSION.'

V3_M

'Monto del ingreso por INDEMNIZACION POR DESPIDO.'

V4_M

'Monto del ingreso por SEGURO DE DESEMPLEO.'

V5_M

'Monto del ingreso por SUBSIDIO O AYUDA SOCIAL (EN DINERO) DEL GOBIERNO, IGLESIAS, ETC.'

V8_M

'Monto del ingreso por ALQUILER (VIVIENDA , TERRENO, OFICINA, ETC.) DE SU PROPIEDAD.'

V9_M

'Monto del ingreso por GANANCIAS DE ALGUN NEGOCIO EN EL QUE NO TRABAJo.'

V10_M

'Monto del ingreso por INTERESES O RENTAS POR PLAZOS FIJOS/INVERSIONES.'

V11_M

'Monto del ingreso por BECA DE ESTUDIO.'

V12_M

'Monto del ingreso por CUOTAS DE ALIMENTOS O AYUDA EN DINERO DE PERSONAS QUE NO VIVEN EN EL HOGAR.'

V18_M

'Monto del ingreso por OTROS INGRESOS EN EFECTIVO (LIMOSNAS, JUEGOS DE AZAR, ETC.).'

V19_AM

'Monto del ingreso por TRABAJO DE MENORES DE 10 ANIOS.'

V21_M

'Monto del ingreso por aguinaldo.'

T_VI

'MONTO TOTAL DE INGRESOS NO LABORALES'

ITF

'MONTO DEL INGRESO TOTAL FAMILIAR.'

DECIFR

'Nro de decil de ingreso total familiar del TOTAL.'

IDECIFR

'Nro de decil de ingreso total familiar del INTERIOR.'

RDECIFR

'Nro de decil de ingreso total familiar de la REGION.'

GDECIFR

'Nro de decil de ingreso total familiar del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES.'

PDECIFR

'Nro de decil de ingreso total familiar del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES.'

ADECIFR

'Nro de decil de ingreso total familiar del AGLOMERADO (ver Anexo I).'

IPCF

'MONTO DEL INGRESO PER CAPITA FAMILIAR.'

DECCFR

'Nro de decil de ingreso per capita familiar del TOTAL EPH (ver Anexo I).'

IDECCFR

'Nro de decil de ingreso per capita familiar del INTERIOR EPH (ver Anexo I).'

RDECCFR

'Nro de decil de ingreso per capita familiar de la REGION.'

GDECCFR

'Nro de decil de ingreso per capita familiar del CONJUNTO DE AGLOMERADOS DE 500 MIL Y MAS HABITANTES.'

PDECCFR

'Nro de decil de ingreso per capita familiar del CONJUNTO DE AGLOMERADOS DE MENOS DE 500 MIL HABITANTES.'

ADECCFR

'Nro de decil de ingreso per capita familiar del AGLOMERADO.'

PONDIH

'Ponderador del ingreso total familiar y del ingreso per capita familiar, para hogares.

...