Calculo del desvío estándar y el coeficiente de variación

Asigna a una estimación de un total poblacional el desvío estándar o el coeficiente de variación correspondiente según las tablas de error muestral de INDEC para EPH continua a partir del segundo trimestre 2003. (Ver errores_muestrales)

calculate_errors(
  value,
  codigo_aglo = "Total",
  periodo_eph = "2014.03",
  measure = "cv"
)

Arguments

value

Vector numérico de las estimaciones de población para las que se desea hallar el desvío estándar o el coeficiente de variación.

codigo_aglo

default = "Total". String con el código numerico del aglomerado al que pertenecen las estimaciones. "Total" para trabajar estimaciones del conjunto de 31 aglomerados urbanos.

periodo_eph

default = "2014.03". String indicando el periodo al que corresponde la EPH. "2014.03" para obtener los errores muestrales correspondientes al tercer trimestre de 2014 en adelante. "2003.03_2014.02" para los errores muestrales del tercer trimestre del 2003 al segundo trimestre del 2014.

measure

default = "cv". String indicando la medida que se desea obtener. "cv" para obtener el coeficiente de variación correspondiente a las estimaciones o "ds" para obtener el desvío estándar.

Details

disclaimer: El script no es un producto oficial de INDEC.

Examples

tabla <- eph::toybase_individual_2016_03 %>%
  eph::organize_labels() %>%
  dplyr::filter(AGLOMERADO == 32) %>%
  calculate_tabulates(x = "CH03",
                         weights = "PONDERA",
                         add.totals = "row")
#> Warning: You are trying to put value labels on factor. It can lead to unexpected results. Factor will be converted to character.
#> Warning: You are trying to put value labels on factor. It can lead to unexpected results. Factor will be converted to character.
tabla %>%
  dplyr::mutate(ds = calculate_errors(Freq, codigo_aglo = "32",
   periodo_eph = "2014.03", measure = "ds"))
#>               CH03   Freq   ds
#>             Jefe/a  38443 5869
#>   Conyuge / Pareja  26817 4990
#>  Hijo/a Hijastro/a  30886 5303
#>           Nietro/a   7962 4282
#>        Madre/Padre   2838 4282
#>          Hermano/a   3619 4282
#>   Otros Familiares   2392 4282
#>      No familiares   1648 4282
#>              Total 114605 9560