Calculo de Pobreza e Indigencia

Funcion para calcular la pobreza e indigencia siguiendo la metodologia de linea.

calculate_poverty(base, basket, print_summary = TRUE)

Arguments

base
Base individual de uno o mas periodos

basket
canasta basica alimentaria y total, con la estructura de la canasta de ejemplo (ver canastas_reg_example)

print_summary
default = TRUE. Opcion para imprimir las tasas de pobreza e indigencia (proporcion de individuos)

Details

disclaimer: El script no es un producto oficial de INDEC.

Examples

bases <- dplyr::bind_rows(toybase_individual_2016_03,
                          toybase_individual_2016_04)

base_pobreza <- calculate_poverty(base = bases,
                                  basket = canastas_reg_example,
                                  print_summary=TRUE)

#> # A tibble: 2 x 4
#> # Groups:   ANO4 [1]
#>    ANO4 TRIMESTRE Tasa_pobreza Tasa_indigencia
#>   <int>     <int>        <dbl>           <dbl>
#> 1  2016         3       0.0296          0.0136
#> 2  2016         4       0.0324          0.0122