Seleccion aleatoria de casos de la base 2016 trimestre 4 para la base hogar

Seleccion aleatoria de casos de la base 2016 trimestre 4 para la base hogar

toybase_hogar_2016_04

Formato

Un data frame con 2000 filas y 177 variables

CODUSU

'Codigo para distinguir VIVIENDAS, permite aparearlas con Hogares y Personas. Ademas permite hacer el seguimiento a traves de los trimestres'

ANO4

'Anio de relevamiento'

TRIMESTRE

'Ventana de Observacion'

NRO_HOGAR

'Codigo para distinguir HOGARES, permite aparearlos con Personas'

REALIZADA

'Entrevista individual realizada'

REGION

'Codigo de Region'

MAS_500

'Aglomerados segun tamanio'

AGLOMERADO

'Codigo de Aglomerado'

PONDERA

'Ponderacion'

IV1

'Tipo de vivienda'

IV1_ESP

'Otros (Especificar)'

IV2

'Cuantos ambientes/ habitaciones tiene la vivienda en total? (sin contar banio/s, cocina, pasillo/s,lavadero, garage)'

IV3

'Los pisos interiores son principalmente de...'

IV3_ESP

'Otros. Especificar'

IV4

'La cubierta exterior del techo es de....'

IV5

'El techo tiene cielorraso/revestimiento interior?'

IV6

'Tiene agua...'

IV7

'El agua es de...'

IV7_ESP

'Otras fuentes. Especificar'

IV8

'Tiene banio/letrina?'

IV9

'El banio o letrina esta...'

IV10

'El banio tiene...'

IV11

'El desague del banio es...'

IV12_1

'La vivienda esta ubicada cerca de basural/es(3 cuadras O menos)'

IV12_2

'La vivienda esta ubicada en zona inundable (en los ultimos 12 meses)'

IV12_3

'La vivienda esta ubicada en villa de emergencia (por observacion)'

II1

'Cuantos ambientes/ habitaciones tiene este hogar para su uso exclusivo?'

II2

'De esos, cuantos usan habitualmente para dormir?'

II3

'Utiliza alguno exclusivamente como lugar de Trabajo(paraconsultorio,estudio,taller, negocio, etc.).'

II3_1

'Si utiliza alguno exclusivamente como lugar de Trabajo Cuantos?.'

II4_1

'Tiene ademas... Cuarto de cocina.'

II4_2

'Tiene ademas... Lavadero.'

II4_3

'Tiene ademas... Garage.'

II5

'De esos ...(los si de preg.4) usan alguno para dormir.'

II5_1

'Si Utiliza alguno para dormir Cuantos?.'

II6

'De esos ...(los si de preg.4) utiliza alguno de estos exclusivamente como lugar de trabajo (consultorio, estudio, taller, negocio, etc.).'

II6_1

'Si utiliza alguno exclusivamente como lugar de trabajo Cuantos?'

II7

'Regimen de tenencia'

II7_ESP

'Otra situacion (Especificar).'

II8

'Combustible utilizado para cocinar'

II8_ESP

'Otro. Especificar.'

II9

'Banio (tenencia y uso).'

V1

'En los ultimos tres meses las personas de este hogar han vivido... de lo que ganan en el trabajo?.'

V2

'En los ultimos tres meses las personas de este hogar han vivido... de alguna jubilacion o pension?.'

V21

'En los ultimos tres meses las personas de este hogar han vivido de... Aguinaldo de alguna jubilacion o pension cobrada el mes anterior?'

V22

'En los ultimos tres meses las personas de este hogar han vivido de... Retroactivo de alguna jubilacion o pension que cobro el mes anterior?'

V3

'En los ultimos tres meses las personas de este hogar han vivido de... indemnizacion por despido?'

V4

'En los ultimos tres meses las personas de este hogar han vivido de... seguro de desempleo?'

V5

'En los ultimos tres meses las personas de este hogar han vivido de... desubsidio oayuda social(en dinero) del gobierno, iglesias, etc.?'

V6

'En los ultimos tres meses las personas de este hogar han vivido con... mercaderias, ropa, alimentos del gobierno iglesias, escuelas, etc.?'

V7

'En los ultimos tres meses las personas de este hogar han vivido con... mercaderias, ropa, alimentos de familiares, vecinos u otras personas que no viven en este hogar? '

V8

'En los ultimos tres meses las personas de este hogar han vivido de... algun alquiler (por una vivienda, terreno, oficina, etc.)de su propiedad?'

V9

'En los ultimos tres meses las personas de este hogar han vivido de... ganancias de algun negocio enel que no trabajan?'

V10

'En los ultimos tres meses las personas de este hogar han vivido de... intereses o rentas por plazos fijos/ inversiones?'

V11

'En los ultimos tres meses las personas de este hogar han vivido de... una beca de estudio?'

V12

'En los ultimos tres meses las personas de este hogar han vivido de... cuotas de alimentos o ayuda en dinero de personas que no viven en el hogar?'

V13

'En los ultimos tres meses las personas de este hogar han vivido con... gastar lo que tenian ahorrado?'

V14

'En los ultimos tres meses las personas de este hogar han vivido de... pedir prestamos a familiares/amigos'

V15

'En los ultimos tres meses las personas de este hogar han vivido de... pedir prestamos a bancos, financieras, etc.?'

V16

'En los ultimos tres meses las personas de este hogar han vivido de... Compran en cuotas o al fiado con tarjeta de credito o libreta?'

V17

'Han tenido que vender alguna de sus pertenencias?'

V18

'Tuvieron otros ingresos en efectivo (limosnas, juegos de azar, etc.)'

V19_A

'menores de 10 anios ayudan con algun dinero trabajando?'

V19_B

'Menores de 10 anios ayudan con algun dinero pidiendo?'

IX_TOT

'Cantidad de miembros del Hogar'

IX_MEN10

'Cantidad de miembros del Hogar menores de 10 anios'

IX_MAYEQ10

'Cantidad de miembros del Hogar de 10 y mas anios'

ITF

'MONTO DE INGRESO TOTAL FAMILIAR'

DECIFR

'Nro de decil del ingreso total del hogar del TOTAL EPH'

IDECIFR

'Nro de decil del ingreso total del hogar del TOTAL EPH'

RDECIFR

'Nro de decil de ingreso total del hogar de la REGION.'

GDECIFR

'Nro de decil de ingreso total del hogar del Conjunto de AGLOMERADOS DE 500 MIL Y MAS HABITANTES.'

PDECIFR

'Nro de decil de ingreso total del hogar del Conjunto de AGLOMERADOS DE MENOS DE 500 MIL HABITANTES'

ADECIFR

'Nro de decil de ingreso total del hogar del AGLOMERADO'

IPCF

'Monto de ingreso per capita familiar(ver AnexoI)'

DECCFR

'Nro de decil del ingreso per capita familiar del TOTAL EPH'

IDECCFR

'Nro de decil del ingreso per capita familiar del INTERIOR'

RDECCFR

'Nro de decil de ingreso per capita familiar de la REGION'

GDECCFR

'Nro de decil de ingreso per capita familiar del Conjunto de AGLOMERADOS DE 500 MIL Y MAS HABITANTES'

PDECCFR

'Nro de decil de ingreso per capita familiar del Conjunto de AGLOMERADOS DE MENOS DE 500 MIL HABITANTES'

ADECCFR

'Nro de decil de ingreso per capita familiar del AGLOMERADO'

PONDIH

'Ponderador del ingreso total familiar y del ingreso per capita familiar'

VII1_1

'Realizacion de las tareas de la casa - Numero de componente del hogar'

VII1_2

'Realizacion de las tareas de la casa - Numero de componente del hogar'

VII2_1

'Otras personas que ayudan en las tareas de la casa - Numero de componente del hogar'

VII2_2

'Otras personas que ayudan en las tareas de la casa - Numero de componente del hogar'

VII2_3

'Otras personas que ayudan en las tareas de la casa - Numero de componente del hogar'

VII2_4

'Otras personas que ayudan en las tareas de la casa - Numero de componente del hogar'

...